Address PR review feedback
- Use shared localisation helpers for wizard fixture maps.
- Ensure wizard route aria-label follows localised route names.
- Add walk completion moment and stat translations across locales.
- Split large fixture modules to keep files <400 lines.
