+++ src/app/data/fixture-localization.ts
@@ -80,11 +80,15 @@ export const localisation = (
  * Most locales use the original brand name, but RTL and non-Latin scripts may need overrides.
  *
  * @param latinName - The brand name in Latin script (e.g., "Facebook")
- * @param scriptOverrides - Names for locales using non-Latin scripts (ar, he, hi, ko, ta)
+ * @param scriptOverrides - Names for locales using non-Latin scripts (ar, he, hi, ko, ta).
+ *   This intentionally targets the scripts where we most commonly transliterate brand names in the UI; most other
+ *   supported locales keep the Latin spelling rather than introducing a locale-specific translation.
  * @param context - Optional context string for debugging
  */
 export const brandLocalizations = (
   latinName: string,
+  // Only RTL and select Indic/CJK scripts typically transliterate brand names; Japanese, Chinese, Thai, Greek, and
+  // Cyrillic usually retain the Latin spellings shown in apps and maps.
   scriptOverrides: Partial<Record<"ar" | "he" | "hi" | "ko" | "ta", string>> = {},
   context?: string,
 ): EntityLocalizations => {
