+++ src/app/data/walk-complete-moment-localizations.ts
@@ -1,0 +1,121 @@
+/** @file Localisation maps for walk completion moments. */
+
+import type { EntityLocalizations } from "../domain/entities/localization";
+import { fillLocalizations } from "./fixture-localization";
+
+export const coffeeMomentLocalizations: EntityLocalizations = fillLocalizations(
+  {
+    "en-GB": { name: "Blue Bottle Coffee", description: "Perfect cortado & friendly barista" },
+    "en-US": { name: "Blue Bottle Coffee", description: "Perfect cortado & friendly barista" },
+    ar: { name: "بلو بوتل كوفي", description: "كورتادو مثالي وباريستا ودود" },
+    cy: { name: "Blue Bottle Coffee", description: "Cortado perffaith a barista cyfeillgar" },
+    da: { name: "Blue Bottle Coffee", description: "Perfekt cortado og venlig barista" },
+    de: { name: "Blue Bottle Coffee", description: "Perfekter Cortado und freundlicher Barista" },
+    el: { name: "Blue Bottle Coffee", description: "Τέλειο cortado και φιλικός barista" },
+    es: { name: "Blue Bottle Coffee", description: "Cortado perfecto y barista amable" },
+    fi: { name: "Blue Bottle Coffee", description: "Täydellinen cortado ja ystävällinen barista" },
+    fr: { name: "Blue Bottle Coffee", description: "Cortado parfait et barista accueillant" },
+    gd: { name: "Blue Bottle Coffee", description: "Cortado foirfe agus barista càirdeil" },
+    he: { name: "בלו בוטל קפה", description: "קורטאדו מושלם ובריסטה ידידותי" },
+    hi: { name: "ब्लू बॉटल कॉफी", description: "परफ़ेक्ट कॉर्टाडो और दोस्ताना बरिस्ता" },
+    it: { name: "Blue Bottle Coffee", description: "Cortado perfetto e barista cordiale" },
+    ja: { name: "ブルーボトルコーヒー", description: "完璧なコルタードと親切なバリスタ" },
+    ko: { name: "블루보틀 커피", description: "완벽한 코르타도와 친절한 바리스타" },
+    nb: { name: "Blue Bottle Coffee", description: "Perfekt cortado og hyggelig barista" },
+    nl: { name: "Blue Bottle Coffee", description: "Perfecte cortado en vriendelijke barista" },
+    pl: { name: "Blue Bottle Coffee", description: "Idealne cortado i sympatyczny barista" },
+    pt: { name: "Blue Bottle Coffee", description: "Cortado perfeito e barista simpático" },
+    ru: { name: "Blue Bottle Coffee", description: "Идеальный кортадо и дружелюбный бариста" },
+    sv: { name: "Blue Bottle Coffee", description: "Perfekt cortado och vänlig barista" },
+    ta: { name: "ப்ளூ பாட்டில் காபி", description: "சிறந்த கொர்டாடோ மற்றும் நட்பான பாரிஸ்டா" },
+    th: { name: "Blue Bottle Coffee", description: "คอร์ตาโดเยี่ยมและบาริสต้าเป็นกันเอง" },
+    tr: { name: "Blue Bottle Coffee", description: "Mükemmel cortado ve güler yüzlü barista" },
+    vi: { name: "Blue Bottle Coffee", description: "Cortado hoàn hảo và barista thân thiện" },
+    "zh-CN": { name: "蓝瓶咖啡", description: "完美的科尔塔多与友好的咖啡师" },
+    "zh-TW": { name: "藍瓶咖啡", description: "完美的科爾塔多與親切的咖啡師" },
+  },
+  "en-GB",
+  "moment: coffee",
+);
+
+export const muralMomentLocalizations: EntityLocalizations = fillLocalizations(
+  {
+    "en-GB": { name: "Hidden Mural", description: "Amazing street art in quiet alley" },
+    "en-US": { name: "Hidden Mural", description: "Amazing street art in quiet alley" },
+    ar: { name: "جدارية مخفية", description: "فن شارع مذهل في زقاق هادئ" },
+    cy: { name: "Murlun Cudd", description: "Celf stryd anhygoel mewn lôn dawel" },
+    da: { name: "Skjult vægmaleri", description: "Fantastisk gadekunst i en stille gyde" },
+    de: {
+      name: "Verstecktes Wandbild",
+      description: "Großartige Street Art in einer ruhigen Gasse",
+    },
+    el: { name: "Κρυφή τοιχογραφία", description: "Εντυπωσιακή street art σε ήσυχο σοκάκι" },
+    es: { name: "Mural oculto", description: "Arte urbano increíble en un callejón tranquilo" },
+    fi: { name: "Piilotettu muraali", description: "Upeaa katutaidetta hiljaisessa kujassa" },
+    fr: { name: "Fresque cachée", description: "Superbe street art dans une ruelle calme" },
+    gd: {
+      name: "Mural falaichte",
+      description: "Ealain sràide iongantach ann an caol-shràid shàmhach",
+    },
+    he: { name: "ציור קיר נסתר", description: "אמנות רחוב מדהימה בסמטה שקטה" },
+    hi: { name: "छिपा हुआ भित्तिचित्र", description: "शांत गली में शानदार स्ट्रीट आर्ट" },
+    it: { name: "Murale nascosto", description: "Street art incredibile in un vicolo tranquillo" },
+    ja: { name: "隠れた壁画", description: "静かな路地にある素晴らしいストリートアート" },
+    ko: { name: "숨겨진 벽화", description: "조용한 골목의 멋진 스트리트 아트" },
+    nb: { name: "Skjult veggmaleri", description: "Fantastisk gatekunst i en rolig bakgate" },
+    nl: {
+      name: "Verborgen muurschildering",
+      description: "Geweldige street art in een rustige steeg",
+    },
+    pl: { name: "Ukryty mural", description: "Niesamowite street art w cichej alejce" },
+    pt: { name: "Mural escondido", description: "Arte urbana incrível num beco tranquilo" },
+    ru: { name: "Скрытая фреска", description: "Потрясающий стрит-арт в тихом переулке" },
+    sv: { name: "Gömd mural", description: "Fantastisk street art i en lugn gränd" },
+    ta: { name: "மறைந்த சுவர் ஓவியம்", description: "அமைதியான வழித்தெருவில் அற்புதமான தெருக் கலை" },
+    th: { name: "ภาพจิตรกรรมฝาผนังที่ซ่อนอยู่", description: "สตรีทอาร์ตสุดเจ๋งในตรอกเงียบ ๆ" },
+    tr: { name: "Gizli duvar resmi", description: "Sakin bir ara sokakta harika sokak sanatı" },
+    vi: { name: "Bích hoạ ẩn", description: "Street art tuyệt vời trong con hẻm yên tĩnh" },
+    "zh-CN": { name: "隐藏壁画", description: "安静小巷里的惊艳街头艺术" },
+    "zh-TW": { name: "隱藏壁畫", description: "安靜巷弄裡的驚豔街頭藝術" },
+  },
+  "en-GB",
+  "moment: mural",
+);
+
+export const parkMomentLocalizations: EntityLocalizations = fillLocalizations(
+  {
+    "en-GB": { name: "Meridian Hill Park", description: "Peaceful pond with friendly ducks" },
+    "en-US": { name: "Meridian Hill Park", description: "Peaceful pond with friendly ducks" },
+    ar: { name: "حديقة ميريديان هيل", description: "بركة هادئة وبط ودود" },
+    cy: { name: "Meridian Hill Park", description: "Pwll tawel gyda hwyaid cyfeillgar" },
+    da: { name: "Meridian Hill Park", description: "Rolig dam med venlige ænder" },
+    de: { name: "Meridian Hill Park", description: "Ruhiger Teich mit freundlichen Enten" },
+    el: { name: "Meridian Hill Park", description: "Ήρεμη λιμνούλα με φιλικές πάπιες" },
+    es: { name: "Meridian Hill Park", description: "Estanque tranquilo con patos amistosos" },
+    fi: { name: "Meridian Hill Park", description: "Rauhallinen lampi ja ystävällisiä ankkoja" },
+    fr: { name: "Meridian Hill Park", description: "Étang paisible avec des canards amicaux" },
+    gd: { name: "Meridian Hill Park", description: "Lochan sìtheil le tunnagan càirdeil" },
+    he: { name: "פארק מרידיאן היל", description: "בריכה שלווה עם ברווזים ידידותיים" },
+    hi: { name: "मेरिडियन हिल पार्क", description: "शांत तालाब और दोस्ताना बत्तखें" },
+    it: { name: "Meridian Hill Park", description: "Laghetto tranquillo con anatre amichevoli" },
+    ja: { name: "メリディアン・ヒル・パーク", description: "穏やかな池と人懐こいアヒル" },
+    ko: { name: "메리디언 힐 파크", description: "평화로운 연못과 친근한 오리" },
+    nb: { name: "Meridian Hill Park", description: "Rolig dam med vennlige ender" },
+    nl: { name: "Meridian Hill Park", description: "Rustige vijver met vriendelijke eenden" },
+    pl: { name: "Meridian Hill Park", description: "Spokojny staw z przyjaznymi kaczkami" },
+    pt: { name: "Meridian Hill Park", description: "Lago tranquilo com patos simpáticos" },
+    ru: { name: "Парк Меридиан-Хилл", description: "Тихий пруд с дружелюбными утками" },
+    sv: { name: "Meridian Hill Park", description: "Lugn damm med vänliga änder" },
+    ta: { name: "மெரிடியன் ஹில் பார்க்", description: "அமைதியான குளமும் நட்பான வாத்துகளும்" },
+    th: { name: "Meridian Hill Park", description: "สระน้ำเงียบสงบกับเป็ดน่ารัก" },
+    tr: { name: "Meridian Hill Park", description: "Sakin bir gölet ve sevimli ördekler" },
+    vi: {
+      name: "Meridian Hill Park",
+      description: "Hồ nước yên bình với những chú vịt thân thiện",
+    },
+    "zh-CN": { name: "梅里迪安山公园", description: "宁静的池塘与亲人的鸭子" },
+    "zh-TW": { name: "梅里迪安山公園", description: "寧靜的池塘與親人的鴨子" },
+  },
+  "en-GB",
+  "moment: park",
+);
