+++ src/app/data/wizard-route-localizations.ts
@@ -1,0 +1,179 @@
+/** @file Wizard route summary localisation maps. */
+
+import type { EntityLocalizations } from "../domain/entities/localization";
+import { localisation } from "./fixture-localization";
+
+// Route Summary Localizations
+// =============================================================================
+
+export const routeTitleLocalizations: EntityLocalizations = localisation(
+  {
+    name: "Hidden Gems Loop",
+    description:
+      "A personalised walk blending street-art alleyways, independent cafés, and serene waterfront views.",
+  },
+  {
+    ar: {
+      name: "جولة الجواهر المخفية",
+      description:
+        "نزهة مخصصة تمزج أزقة فن الشارع والمقاهي المستقلة وإطلالات الواجهة البحرية الهادئة.",
+    },
+    cy: {
+      name: "Llwybr Trysorau Cudd",
+      description:
+        "Taith bersonol sy'n cyfuno strydoedd celf, caffis annibynnol, a golygfeydd tawel o'r lan.",
+    },
+    da: {
+      name: "Skjulte Perler-sløjfen",
+      description:
+        "En personlig gåtur, der blander street art-gyder, uafhængige caféer og rolige udsigter ved havnefronten.",
+    },
+    de: {
+      name: "Verborgene Schätze Runde",
+      description:
+        "Ein personalisierter Spaziergang, der Street-Art-Gassen, unabhängige Cafés und ruhige Aussichtspunkte am Wasser verbindet.",
+    },
+    el: {
+      name: "Κύκλος Κρυμμένων Θησαυρών",
+      description:
+        "Μια εξατομικευμένη βόλτα που συνδυάζει σοκάκια street art, ανεξάρτητα καφέ και ήρεμες θέες στο νερό.",
+    },
+    es: {
+      name: "Circuito de Joyas Ocultas",
+      description:
+        "Una caminata personalizada que mezcla callejones de arte urbano, cafeterías independientes y miradores tranquilos junto al agua.",
+    },
+    fi: {
+      name: "Kätkettyjen helmien kierros",
+      description:
+        "Räätälöity kävely, jossa yhdistyvät street art -kujat, itsenäiset kahvilat ja rauhalliset rantamaisemat.",
+    },
+    fr: {
+      name: "Boucle des Trésors Cachés",
+      description:
+        "Une balade personnalisée mêlant ruelles street art, cafés indépendants et points de vue paisibles sur le front de mer.",
+    },
+    gd: {
+      name: "Cuairt nan Ulaidhean Falaichte",
+      description:
+        "Cuairt phearsanta a tha a' cothlamadh sràidean ealain, càbaidean neo-eisimeileach agus seallaidhean sàmhach air an oirthir.",
+    },
+    he: {
+      name: "לולאת האוצרות הנסתרים",
+      description:
+        "מסלול מותאם אישי המשולב בסמטאות אמנות רחוב, בתי קפה עצמאיים ונקודות תצפית שקטות על קו החוף.",
+    },
+    hi: {
+      name: "छिपे रत्नों की लूप",
+      description: "स्ट्रीट आर्ट की गलियों, स्वतंत्र कैफ़े और शांत जल-किनारे दृश्यों का मेल करती एक निजी वॉक।",
+    },
+    it: {
+      name: "Circuito dei Tesori Nascosti",
+      description:
+        "Passeggiata personalizzata che abbina vicoli di street art, caffetterie indipendenti e viste serene sul lungomare.",
+    },
+    ja: {
+      name: "隠れた名所ループ",
+      description:
+        "ストリートアートの路地、個人経営のカフェ、静かなウォーターフロントを組み合わせたカスタムウォークです。",
+    },
+    ko: {
+      name: "숨겨진 보석 루프",
+      description:
+        "스트리트 아트 골목, 독립 카페, 잔잔한 워터프런트 전망을 조합한 맞춤 워크입니다.",
+    },
+    nb: {
+      name: "Skjulte perler-sløyfen",
+      description:
+        "En personlig gåtur som kombinerer street art-gater, uavhengige kafeer og rolige utsiktspunkter ved vannet.",
+    },
+    nl: {
+      name: "Verborgen Parels-lus",
+      description:
+        "Een persoonlijke wandeling die street-artsteegjes, onafhankelijke cafés en rustige uitzichten langs het water combineert.",
+    },
+    pl: {
+      name: "Pętla Ukrytych Pereł",
+      description:
+        "Personalizowany spacer łączący uliczne murale, niezależne kawiarnie i spokojne widoki nad wodą.",
+    },
+    pt: {
+      name: "Rota Joias Escondidas",
+      description:
+        "Uma caminhada personalizada que mistura ruelas de arte urbana, cafés independentes e vistas tranquilas junto à água.",
+    },
+    ru: {
+      name: "Цикл Скрытых Жемчужин",
+      description:
+        "Персональная прогулка, сочетающая переулки со стрит-артом, независимые кафе и тихие виды на набережную.",
+    },
+    sv: {
+      name: "Slingan Dolda Pärlor",
+      description:
+        "En personlig promenad som blandar street‑artgränder, fristående kaféer och lugna utsikter vid vattnet.",
+    },
+    ta: {
+      name: "மறைக்கப்பட்ட நகைகள் வளையம்",
+      description:
+        "தெரு கலை பாதைகள், சுயாதீனக் கஃபேக்கள், அமைதியான நீர்கரை காட்சிகள் கலந்து உருவான தனிப்பயன் நடைபாதை.",
+    },
+    th: {
+      name: "ลูปอัญมณีที่ซ่อนอยู่",
+      description: "เส้นทางแบบเฉพาะตัวที่ผสมตรอกศิลปะข้างถนน คาเฟ่อิสระ และจุดชมวิวริมน้ำอันเงียบสงบ",
+    },
+    tr: {
+      name: "Gizli Mücevherler Döngüsü",
+      description:
+        "Sokak sanatı ara sokakları, bağımsız kafeler ve sakin sahil manzaralarını harmanlayan kişiselleştirilmiş bir yürüyüş.",
+    },
+    vi: {
+      name: "Vòng Viên Ngọc Ẩn",
+      description:
+        "Hành trình cá nhân hóa kết hợp các con hẻm nghệ thuật đường phố, quán cà phê độc lập và điểm ngắm ven sông yên tĩnh.",
+    },
+    "zh-CN": {
+      name: "隐藏宝藏环线",
+      description: "个性化步行体验，融合街头艺术小巷、独立咖啡店与静谧水岸观景点。",
+    },
+    "zh-TW": {
+      name: "隱藏寶藏環線",
+      description: "將街頭藝術巷弄、獨立咖啡館與寧靜水岸景點融合的個人化路線。",
+    },
+  },
+  "route: title",
+);
+
+export const routeBadgeLocalizations: EntityLocalizations = localisation(
+  { name: "Custom" },
+  {
+    ar: { name: "مخصص" },
+    cy: { name: "Personol" },
+    da: { name: "Tilpasset" },
+    de: { name: "Individuell" },
+    el: { name: "Προσαρμοσμένη" },
+    es: { name: "Personalizada" },
+    fi: { name: "Mukautettu" },
+    fr: { name: "Personnalisée" },
+    gd: { name: "Gnàthaichte" },
+    he: { name: "מותאם אישית" },
+    hi: { name: "व्यक्तिगत" },
+    it: { name: "Personalizzato" },
+    ja: { name: "カスタム" },
+    ko: { name: "맞춤형" },
+    nb: { name: "Tilpasset" },
+    nl: { name: "Aangepast" },
+    pl: { name: "Indywidualny" },
+    pt: { name: "Personalizada" },
+    ru: { name: "Индивидуальный" },
+    sv: { name: "Anpassad" },
+    ta: { name: "தனிப்பயன்" },
+    th: { name: "ปรับแต่ง" },
+    tr: { name: "Özel" },
+    vi: { name: "Tùy chỉnh" },
+    "zh-CN": { name: "自定义" },
+    "zh-TW": { name: "自訂" },
+  },
+  "route: badge",
+);
+
+// =============================================================================
