+++ tests/walk-completion-data.test.ts
@@ -10,6 +10,7 @@ import {
   walkCompletionShareOptions,
 } from "../src/app/data/stage-four";
 import { pickLocalization } from "../src/app/domain/entities/localization";
+import { SUPPORTED_LOCALES } from "../src/app/i18n/supported-locales";
 
 /**
  * Helper to test stat localizations across multiple locales.
@@ -89,10 +90,22 @@ describe("walkCompletionMoments localizations", () => {
     });
   });
 
+  it("moments provide localised name and description across all supported locales", () => {
+    expect(walkCompletionMoments.length).toBeGreaterThan(0);
+
+    for (const locale of SUPPORTED_LOCALES) {
+      for (const moment of walkCompletionMoments) {
+        const localized = pickLocalization(moment.localizations, locale.code);
+        expect(localized.name).toBeTruthy();
+        expect(localized.description).toBeTruthy();
+      }
+    }
+  });
+
   it("coffee moment has correct English localization", () => {
     const coffeeMoment = walkCompletionMoments.find((m) => m.id === "coffee");
     expect(coffeeMoment).toBeDefined();
-    if (!coffeeMoment) return;
+    if (!coffeeMoment) throw new Error("Missing moment: coffee");
 
     const localized = pickLocalization(coffeeMoment.localizations, "en-GB");
     expect(localized.name).toBe("Blue Bottle Coffee");
@@ -112,7 +125,7 @@ describe("walkCompletionMoments localizations", () => {
   it("mural moment has correct English localization", () => {
     const muralMoment = walkCompletionMoments.find((m) => m.id === "mural");
     expect(muralMoment).toBeDefined();
-    if (!muralMoment) return;
+    if (!muralMoment) throw new Error("Missing moment: mural");
 
     const localized = pickLocalization(muralMoment.localizations, "en-GB");
     expect(localized.name).toBe("Hidden Mural");
@@ -130,7 +143,7 @@ describe("walkCompletionShareOptions localizations", () => {
   it("Facebook has localized names for non-Latin scripts", () => {
     const facebookOption = walkCompletionShareOptions.find((o) => o.id === "facebook");
     expect(facebookOption).toBeDefined();
-    if (!facebookOption) return;
+    if (!facebookOption) throw new Error("Missing share option: facebook");
 
     // Latin scripts should use "Facebook"
     expect(pickLocalization(facebookOption.localizations, "en-GB").name).toBe("Facebook");
@@ -146,7 +159,7 @@ describe("walkCompletionShareOptions localizations", () => {
   it("Instagram has localized names for non-Latin scripts", () => {
     const instagramOption = walkCompletionShareOptions.find((o) => o.id === "instagram");
     expect(instagramOption).toBeDefined();
-    if (!instagramOption) return;
+    if (!instagramOption) throw new Error("Missing share option: instagram");
 
     // Latin scripts should use "Instagram"
     expect(pickLocalization(instagramOption.localizations, "en-GB").name).toBe("Instagram");
